refactor: update model name generator imports to use consolidated nameGenerator module
